Remove authority selection from onboarding (#679)
