+++ app/Entry.js
@@ -19,7 +19,6 @@ import Onboarding2 from './views/onboarding/Onboarding2';
 import Onboarding3 from './views/onboarding/Onboarding3';
 import Onboarding4 from './views/onboarding/Onboarding4';
 import Onboarding5 from './views/onboarding/Onboarding5';
-import Onboarding6 from './views/onboarding/Onboarding6';
 import { SettingsScreen } from './views/Settings';
 
 const Stack = createStackNavigator();
@@ -90,11 +89,6 @@ class Entry extends Component {
           <Stack.Screen
             name='Onboarding5'
             component={Onboarding5}
-            options={{ headerShown: false }}
-          />
-          <Stack.Screen
-            name='Onboarding6'
-            component={Onboarding6}
             options={{ headerShown: false }}
           />
           <Stack.Screen
